+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Non-trivial `inherited` expressions are excluded in `RedundantParentheses`.
+
 ## [1.9.0] - 2024-09-03
 
 ### Added