+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.10.0] - 2024-10-01
+
 ### Added
 
 - `FullyQualifiedImportCheck` analysis rule, which flags non-fully qualified imports.
@@ -1369,7 +1371,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.9.0...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.10.0...HEAD
+[1.10.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v1.6.0...v1.7.0